Guess the number terminal based game
Python code for a terminal based  guess the number game. This will generate a random number between 1 to 50 and the player has to guess the number. If the player answers wrong the game will provide a hint whether the answer is smaller or greater than the number player guessed. It goes on for total 10 rounds.